move additionalUnit in service
